Trait lexical_parse_integer::FromLexical

pub trait FromLexical: Number {
    // Required methods
;
;
}

Trait for numerical types that can be parsed from bytes.

Required Methods

Checked parser for a string-to-number conversion.

This method parses the entire string, returning an error if any invalid digits are found during parsing. Returns a Result containing either the parsed value, or an error containing any errors that occurred during parsing.

bytes - Slice containing a numeric string.

Checked parser for a string-to-number conversion.

This method parses until an invalid digit is found (or the end of the string), returning the number of processed digits and the parsed value until that point. Returns a Result containing either the parsed value and the number of processed digits, or an error containing any errors that occurred during parsing.

bytes - Slice containing a numeric string.